refactor(prover): address commitment backend review
