+++ crates/akita-prover/src/lib.rs
@@ -38,24 +38,24 @@ pub use commitment::{
     CommitmentExecutor, CommitmentExecutorBuilder, CommitmentNttRequirement, CommitmentNttRoute,
     CommitmentNttStage, CommitmentOperationContext, CommitmentOperationId,
     CommitmentRequestCapabilities, CommitmentResourceControl, CommitmentRoundStep,
-    CommitmentSource, CommitmentStateBinding, CommitmentStateComponents, CommitmentStateOutput,
-    CommitmentStatePolicy, CompiledCommitmentRequest, CompressionOperation,
-    CompressionOperationCapabilities, CompressionStageOutput, CompressionState,
-    DenseCoefficientSource, DenseRepresentation, DenseType, ExternalFusedInnerCommitmentEncoder,
-    ExternalInnerCommitmentCapability, ExternalInnerCommitmentInput,
-    ExternalInnerCommitmentOperation, ExternalOperationIdentity, FullCommitmentOutput,
-    FusedInnerOuterOperation, GroupContext, InnerCommitOperation, InnerCommitOutput, InnerImage,
-    InnerImageExportOperation, InnerImageInput, InnerOuterRouteKind, InnerRelationState,
-    InnerRelationStateMaterial, IntoPortableCommitmentState, NoRetainedStatePolicy,
-    OneHotIndexWidth, OneHotRepresentation, OneHotType, OuterCommitOperation, OuterCommitPlan,
-    OuterCompressionState, PolynomialRepresentation, PolynomialType, PolynomialTypeSelection,
-    PortableCommitmentState, PortableCompressionState, PortableCompressionStateExport,
-    PortableStatePolicy, PredecomposedDigitPlanes, PreparedCommitmentResources,
-    PreparedCompression, PreparedExternalInnerCommitment, PreparedFusedCommitment,
-    PreparedInnerCommitment, PreparedOuterCommitment, ResidentCommitmentState, ResidentStatePolicy,
-    ResolvedCommitSource, ShortNormRepresentation, ShortNormType, StageDimensionCapabilities,
-    StageResources, StateOwnerCapability, TerminalBindingState, TerminalTFieldsMessage,
-    UncompressedCommitPlan, UncompressedCommitmentOutput, UnitPositionSlice,
+    CommitmentSource, CommitmentStateBinding, CommitmentStateComponents, CommitmentStatePolicy,
+    CompiledCommitmentRequest, CompressionOperation, CompressionOperationCapabilities,
+    CompressionStageOutput, CompressionState, DenseCoefficientSource, DenseRepresentation,
+    DenseType, ExternalFusedInnerCommitmentEncoder, ExternalInnerCommitmentCapability,
+    ExternalInnerCommitmentInput, ExternalInnerCommitmentOperation, ExternalOperationIdentity,
+    FullCommitmentOutput, FusedInnerOuterOperation, GroupContext, InnerCommitOperation,
+    InnerCommitOutput, InnerImage, InnerImageExportOperation, InnerImageInput, InnerOuterRouteKind,
+    InnerRelationState, InnerRelationStateMaterial, IntoPortableCommitmentState,
+    NoRetainedStatePolicy, OneHotIndexWidth, OneHotRepresentation, OneHotType,
+    OuterCommitOperation, OuterCommitPlan, OuterCompressionState, PolynomialRepresentation,
+    PolynomialType, PolynomialTypeSelection, PortableCommitmentState, PortableCompressionState,
+    PortableCompressionStateExport, PortableStatePolicy, PredecomposedDigitPlanes,
+    PreparedCommitmentResources, PreparedCompression, PreparedExternalInnerCommitment,
+    PreparedFusedCommitment, PreparedInnerCommitment, PreparedOuterCommitment,
+    ResidentCommitmentState, ResidentStatePolicy, ResolvedCommitSource, ShortNormRepresentation,
+    ShortNormType, StageDimensionCapabilities, StageResources, StateOwnerCapability,
+    TerminalBindingState, TerminalTFieldsMessage, UncompressedCommitPlan,
+    UncompressedCommitmentOutput, UnitPositionSlice,
 };
 pub use compute::{
     prewarm_ntt_requirements, BatchDecomposeFoldOutcome, CommitCluster, ComputeBackendSetup,
