+++ crates/akita-types/src/lib.rs
@@ -112,9 +112,10 @@ pub use proof::{
     build_compression_relation_weights, build_reduced_compression_relation_weights,
     build_terminal_response, build_terminal_response_from_groups,
     canonical_extension_opening_reduction_shape, canonical_proof_shape,
-    commit_only_setup_field_elements, compression_relation_claim_from_rhs_extension,
-    decode_terminal_z_golomb_payload, derive_public_matrix_prefix, draw_group_fold_challenges,
-    emit_witness_e_planes, emit_witness_r_planes, emit_witness_t_planes, emit_witness_z_planes,
+    commit_only_setup_field_elements, commitment_execution_setup_field_elements,
+    compression_relation_claim_from_rhs_extension, decode_terminal_z_golomb_payload,
+    derive_public_matrix_prefix, draw_group_fold_challenges, emit_witness_e_planes,
+    emit_witness_r_planes, emit_witness_t_planes, emit_witness_z_planes,
     evaluate_reduced_compression_map, folded_root_supports_opening_shape, generate_relation_rhs,
     padded_setup_prefix_len, prepare_coefficient_packing_batch_semantics,
     prepare_coefficient_packing_verifier_batch_semantics, prepare_opening_point,
@@ -136,9 +137,10 @@ pub use proof::{
     CoefficientPackingBatchSemanticInputs, CoefficientPackingBatchSemantics,
     CoefficientPackingChallenges, CoefficientPackingGroupSemantics, CoefficientPackingStage2Source,
     CoefficientPackingStage2Terms, CoefficientPackingVerifierBatchSemantics,
-    CoefficientPackingVerifierGroupSemantics, Commitment, CommitmentVerifier, CommittedGroup,
-    CompressionRelationAddressGeometry, CompressionRelationWeights, DigitBlockIter, DigitBlocks,
-    DummyProof, ExtensionOpeningReductionProof, ExtensionOpeningReductionShape, FoldLevelProof,
+    CoefficientPackingVerifierGroupSemantics, Commitment, CommitmentSetupMatrixShape,
+    CommitmentVerifier, CommittedGroup, CompressionRelationAddressGeometry,
+    CompressionRelationWeights, DigitBlockIter, DigitBlocks, DummyProof,
+    ExtensionOpeningReductionProof, ExtensionOpeningReductionShape, FoldLevelProof,
     GroupBatchStatement, GroupFoldChallenges, LevelProofShape, NegativeBinarySupport,
     NextWitnessBinding, NextWitnessBindingShape, OpeningClaims, OpeningClaimsLayout, OpeningPoints,
     PhysicalL2NormProof, PhysicalResponsePlan, PolynomialGroupClaims, PolynomialGroupLayout,
