+++ crates/akita-types/src/proof/mod.rs
@@ -105,9 +105,10 @@ pub use setup::{
 };
 pub use setup_envelope::{
     accumulate_matrix_field_elements_for_level, accumulate_terminal_matrix_field_elements,
-    commit_only_setup_field_elements, setup_matrix_capacity_for_schedule,
-    setup_matrix_field_elements_for_schedule, setup_prefix_slot_field_elements,
-    verifier_setup_matrix_capacity_for_schedule,
+    commit_only_setup_field_elements, commitment_execution_setup_field_elements,
+    setup_matrix_capacity_for_schedule, setup_matrix_field_elements_for_schedule,
+    setup_prefix_slot_field_elements, verifier_setup_matrix_capacity_for_schedule,
+    CommitmentSetupMatrixShape,
 };
 pub use setup_prefix::{
     active_setup_field_len, padded_setup_prefix_len, scheduled_setup_prefix,
